fix: complete Assets diagnostic propagation
