+++ packages/content-engine/src/query-error.ts
@@ -7,30 +7,6 @@ import {
 } from "./structured-query";
 import { DocumentResolutionLimitError } from "./document-graph/document-resolution";
 import { CachedDocumentLoaderError } from "./document-graph/cached-document-loader";
-import { DocumentSourceCompilationError } from "./document-graph/source-compiler";
-import { MdxDocumentError } from "./mdx";
-import { MarkdownMetadataError } from "./markdown-errors";
-import { MarkdownDocumentError } from "./document-graph/markdown-document";
-
-const findNestedSourceError = (error: unknown) => {
-  const visited = new Set<unknown>();
-  let current = error;
-  let documentError: Error | undefined;
-  while (current !== undefined && visited.has(current) === false) {
-    if (current instanceof MdxDocumentError) {
-      return current;
-    }
-    if (current instanceof MarkdownMetadataError) {
-      return current;
-    }
-    if (current instanceof MarkdownDocumentError) {
-      documentError ??= current;
-    }
-    visited.add(current);
-    current = current instanceof Error ? current.cause : undefined;
-  }
-  return documentError;
-};
 
 export type AssetResourceQueryError = Omit<
   AssetResourceQueryFailure["error"],
@@ -42,8 +18,7 @@ export type AssetResourceQueryError = Omit<
 
 /** Classifies content-query failures independently of an HTTP or RPC transport. */
 export const getAssetResourceQueryError = (
-  error: unknown,
-  { includeSourceDetails = false }: { includeSourceDetails?: boolean } = {}
+  error: unknown
 ): AssetResourceQueryError | undefined => {
   if (error instanceof AssetIndexRevisionError) {
     return {
@@ -82,79 +57,6 @@ export const getAssetResourceQueryError = (
   const visited = new Set<unknown>();
   let cause = error;
   while (cause !== undefined && visited.has(cause) === false) {
-    if (includeSourceDetails && cause instanceof MdxDocumentError) {
-      const details: Record<string, string | number> = {};
-      const position = cause.sourceRange?.start;
-      if (position !== undefined) {
-        details.line = position.line;
-        details.column = position.column;
-      }
-      if (cause.reason !== undefined) {
-        details.reason = cause.reason;
-      }
-      return {
-        code: "INVALID_REQUEST",
-        message: cause.message,
-        retryable: false,
-        details,
-        status: 400,
-      };
-    }
-    if (
-      includeSourceDetails &&
-      cause instanceof DocumentSourceCompilationError &&
-      cause.code === "DOCUMENT_ANALYSIS_FAILED"
-    ) {
-      const sourceError = findNestedSourceError(cause.cause);
-      const details: Record<string, string | number> = {};
-      if (cause.documentId !== undefined) {
-        details.assetId = cause.documentId;
-      }
-      if (cause.documentPath !== undefined) {
-        details.path = cause.documentPath;
-      }
-      if (sourceError instanceof MdxDocumentError) {
-        const position = sourceError.sourceRange?.start;
-        if (position !== undefined) {
-          details.line = position.line;
-          details.column = position.column;
-        }
-        if (sourceError.reason !== undefined) {
-          details.reason = sourceError.reason;
-        }
-        return {
-          code: "INVALID_REQUEST",
-          message: sourceError.message,
-          retryable: false,
-          details,
-          status: 400,
-        };
-      }
-      if (sourceError instanceof MarkdownMetadataError) {
-        if (sourceError.line !== undefined) {
-          details.line = sourceError.line;
-        }
-        if (sourceError.column !== undefined) {
-          details.column = sourceError.column;
-        }
-        return {
-          code: "INVALID_REQUEST",
-          message: sourceError.message,
-          retryable: false,
-          details,
-          status: 400,
-        };
-      }
-      if (sourceError instanceof Error) {
-        return {
-          code: "INVALID_REQUEST",
-          message: sourceError.message,
-          retryable: false,
-          details,
-          status: 400,
-        };
-      }
-    }
     if (cause instanceof DocumentResolutionLimitError) {
       const details: Record<string, string | number> = {};
       for (const [key, value] of Object.entries({
